+++ fits_storage/wsgihandler.py
@@ -35,7 +35,7 @@ from fits_storage.web.fileserver import fileserver, download, download_post
 from fits_storage.web.qastuff import qareport, qametrics, qaforgui
 from fits_storage.web.statistics import content, stats
 from fits_storage.web.user import request_account, password_reset, request_password_reset, login, logout, whoami, \
-    change_password, change_email, admin_change_email, orcid, admin_change_password
+    change_password, change_email, admin_change_email, orcid, admin_change_password, admin_file_permissions
 from fits_storage.web.user import staff_access, user_list
 from fits_storage.web.userprogram import my_programs
 from fits_storage.web.searchform import searchform, nameresolver
@@ -262,7 +262,9 @@ url_map = Map([
     Rule('/curation', curation_report),                             # curation_report handler
     Rule('/staff_access', staff_access),                            # staff_access
     Rule('/admin_change_email', admin_change_email),                # admin page for changing a user's email
-    Rule('/admin_change_password', admin_change_password),             # admin page for changing a user's password
+    Rule('/admin_change_password', admin_change_password),          # admin page for changing a user's password
+    Rule('/admin_file_permissions', admin_file_permissions),        # admin page for changing custom per-file/obsid
+                                                                    #   permissions
 
     Rule('/nameresolver/<resolver>/<target>', nameresolver),        # Name resolver proxy
     Rule('/fileontape/<filename>', fileontape),                     # The fileontape handler